_ARCHIVADO/requirements.txt
opencv-python
dlib
imutils
numpy
Pillow
fer
mediapipe
